fix composite file issue, add proptest
Fix composite file issue. The composite file had an issue with the last written fast field, where the wrong field was set as the last range in the composite file due to sorting.
Fix handling of empty fastfields for ip codec.

+++ src/indexer/merger.rs
@@ -376,7 +376,7 @@ impl IndexMerger {
         let iter = doc_id_mapping.iter().map(|(doc_id, reader_ordinal)| {
             let fast_field_reader = &fast_field_readers[*reader_ordinal as usize];
             fast_field_reader
-                .get((*doc_id) as u64)
+                .get(*doc_id)
                 .unwrap_or(compressor.null_value)
         });
